Reorganize available language check
